requirements.txt
Django==3.0.3
django-jalali
django-mptt
Pillow
psycopg2-binary
djangorestframework
django-cors-headers
django-filter==21.1
django-background-tasks
zeep
kavenegar
